update for dim3 template
